requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
pip==22.2.2
pytz==2023.3
setuptools==62.6.0
sqlparse==0.4.3
django-cors-headers==3.14.0
Pillow==9.5.0